chore: move scripts to tools/, update rating to use media_hash, strip subtitle CSS
