+++ .github/release_notes/v1.02.1.md
@@ -77,7 +77,6 @@ How to reproduce verification locally
   Set-Location 'z:\local-video-server'
   python .\scripts\check_links.py
 
-
 Notes
 
 - If the GitHub CLI is not installed or not authenticated in your environment, you can publish this release manually via the GitHub web UI using tag `v1.02.1` and copying this body into the release notes.
