--- TODO_high.md
@@ -1,37 +1,0 @@
-# Local Video Server v1.02 – Release Notes (November 2025)
-
-## 🧱 Backend
-
-- Replaced legacy `main.py` with optimized, cache-enabled version.
-- Introduced `cache_manager.py` for in-memory caching (95% less I/O).
-- Migrated metadata storage from JSON to SQLite (`video_metadata.db`).
-- Added `/admin/performance` and `/admin/cache/status` monitoring endpoints.
-
-## 🎨 Frontend
-
-- Implemented enhanced cross-platform video preview system.
-- Added VR-compatible long-press previews and touch detection.
-- Improved accessibility and color contrast in Glass+Neo themes.
-- Fixed inconsistent thumbnail loading on hover.
-
-## 🤖 AI Integration
-
-- Updated OpenAI model configuration to `gpt-4o-mini` (200 tokens).
-- Reduced per-chat cost by ~50% while maintaining response quality.
-
-## 📊 Monitoring
-
-- Integrated route performance and cache hit tracking.
-- Enabled real-time performance dashboard under `/admin/performance`.
-
-## 🧪 Quality & Validation
-
-- Verified 60–80% faster response times vs. v1.01.
-- Cache hit rate exceeding 90% in typical use.
-- Video previews stable across desktop, mobile, and VR.
-
-## 🗺 Roadmap Highlights
-
-- Adaptive Bitrate Streaming (ABR) integration (planned v1.1)
-- Redis-based distributed caching (planned v1.1)
-- Enhanced AI metadata search and PWA support (planned v1.2)
