--- copilot_cleanup_todo.md
@@ -1,145 +1,0 @@
-# 🧹 CoPilot TODO.md — Repository Cleanup & Archival Phase
-
-## 🎯 Objective
-Perform a full post‑implementation cleanup of the Local Video Server repository. Remove deprecated, redundant, or unused files (subtitle, legacy, and intermediate documentation) and safely archive them to maintain repo cleanliness and historical traceability.
-
----
-
-## 🗂️ 1. Documentation Cleanup & Archival
-
-### ✅ Tasks Completed
-- All subtitle‑related documentation archived to `docs/deferred/`
-- CoPilot todo tracker updated
-- Active subtitle code removed from runtime
-
-### ⚠️ Tasks Pending
-1. **Identify and archive obsolete documents**:
-   - [ ] Move the following to `docs/deferred/legacy/`:
-     - `SUBTITLE_GENERATION_GUIDE.md` (stub version)
-     - `SUBTITLE_SYSTEM_TROUBLESHOOTING.md` (stub)
-     - `SUBTITLE_GENERATION_COMPLETE.md`
-     - `VIDEO_PREVIEW_FEATURE_OLD.md` (if present)
-     - Any other pre‑optimization notes (verify via grep 'obsolete' / 'legacy')
-   - [ ] Delete empty or placeholder `.md` files that were auto‑generated by prior CoPilot runs and contain fewer than 10 lines.
-   - [ ] Verify all moved documents exist in `docs/deferred/legacy/` and update the index reference file `docs/ARCHIVE_INDEX.md` with their new paths.
-
-### 🧩 Implementation Notes
-- Maintain relative links in deferred docs (update internal `see also:` sections).
-- Keep only active documents: `README.md`, `PERFORMANCE_OPTIMIZATION_GUIDE.md`, `VIDEO_PREVIEW_IMPROVEMENTS.md`, `GLASSMORPHIC_NEOMORPHIC_DESIGN.md`, `ADAPTIVE_STREAMING_SYSTEM.md`.
-
----
-
-## 🐍 2. Python Source File Cleanup
-
-### 🔍 Identification Phase
-1. Run a static analysis to detect unused modules:
-   ```bash
-   pylint --disable=all --enable=unused-import,unused-variable main.py *.py
-   ```
-2. Check for orphan files:
-   - [ ] `subtitles.py`
-   - [ ] `manage_subs.py`
-   - [ ] `subtitle_utils.py`
-   - [ ] `config_subtitles.py`
-   - [ ] Any `test_*.py` not referenced in pytest or workflow configs.
-
-### 🧹 Removal & Archival
-1. **Archive before deletion**:
-   ```bash
-   mkdir -p archive/python_legacy
-   git mv subtitles.py manage_subs.py archive/python_legacy/
-   ```
-2. Update imports in all remaining `.py` files:
-   - [ ] Remove `import subtitles`, `from manage_subs import ...`, and any `SubtitleGenerator` class calls.
-   - [ ] Replace removed functions with `pass` or logging placeholders only where needed for compatibility.
-3. Validate via runtime test:
-   ```bash
-   flask run
-   # or python main.py
-   ```
-4. Update `CHANGELOG.md` with:
-   ```text
-   [1.02] — Subtitle and legacy Python modules removed, docs archived.
-   ```
-
----
-
-## 📁 3. Static Asset Cleanup
-
-### Tasks
-1. Remove legacy subtitle/static assets:
-   - [ ] `static/subtitles/`
-   - [ ] `static/js/subtitle-controls.js`
-   - [ ] `static/css/subtitle.css`
-2. Verify player CSS and JS are intact:
-   - `player-controls.css`
-   - `video-player-controls.js`
-3. Run link validation:
-   ```bash
-   grep -Ri 'subtitle' templates/ static/
-   ```
-   Confirm zero active links remain.
-
----
-
-## 🧠 4. Validation & Testing
-
-### Steps
-1. Run full app smoke‑test:
-   ```bash
-   flask run
-   ```
-   - [ ] Homepage loads correctly
-   - [ ] Watch page functions normally
-   - [ ] No missing asset errors
-   - [ ] No broken imports
-2. Run linters:
-   ```bash
-   pylint main.py
-   black --check .
-   flake8
-   ```
-3. Optional: Run pytest suite if configured:
-   ```bash
-   pytest -q
-   ```
-
----
-
-## 🪶 5. Documentation & Repo Maintenance
-
-1. Update `README.md`:
-   - [ ] Add a short section under **Deprecated Systems** summarizing the archival and cleanup.
-   - [ ] Ensure installation and quick‑start steps remain accurate.
-2. Update `docs/ARCHIVE_INDEX.md` to list all archived files with short summaries.
-3. Remove placeholder `.md` files left by failed patch operations (check `docs/` for <20 line files).
-4. Commit changes:
-   ```bash
-   git add .
-   git commit -m "Cleanup: removed unused subtitle files, archived legacy docs"
-   git push origin main
-   ```
-
----
-
-## 🧾 6. Final Verification Checklist
-
-| Area | Action | Status |
-|------|---------|---------|
-| Documentation | All outdated subtitle docs archived | ☐ |
-| Python Source | Orphan subtitle modules removed | ☐ |
-| Static Assets | Subtitle assets deleted | ☐ |
-| Imports | No references to subtitle modules | ☐ |
-| Linting | Clean run | ☐ |
-| Smoke‑Test | App runs cleanly | ☐ |
-| README | Updated for v1.02 | ☐ |
-
----
-
-## 🧭 Summary
-**Purpose:** Clean the repo of any unused subtitle or legacy components while archiving historical files for traceability.  
-**Outcome:** A minimal, performant, and maintainable Local Video Server repository aligned with the modern architecture and documentation hierarchy.  
-**Owner:** CoPilot  
-**Target Completion:** v1.02 release cycle  
-**Risk Level:** Low — purely maintenance and archival work.
-
