+++ docs/SUBTITLE_GENERATION_GUIDE.md
@@ -5,7 +5,7 @@ The detailed Automatic Subtitle Generation guide was archived and moved to:
 `docs/deferred/legacy/SUBTITLE_GENERATION_GUIDE.md`
 
 If you need the historical implementation notes, configuration examples, or CLI references, open the archived copy in `docs/deferred/legacy/`.
+
 - **Local AI transcription** using OpenAI's Whisper models via faster-whisper
 
 - **WebVTT and SRT format support** for broad compatibility
-
