+++ docs/releases/pr_announcement_v1.02.1.md
@@ -1,6 +1,6 @@
 Title: Local Video Server v1.02.1 Released — Documentation Integrity Audit Complete
 
-Release: https://github.com/russell-henderson/local-server-video/releases/tag/v1.02.1
+Release: <https://github.com/russell-henderson/local-server-video/releases/tag/v1.02.1>
 
 Summary: Patch release verifying documentation integrity; no runtime changes.
