+++ templates/_base.html
@@ -14,10 +14,8 @@
     <!-- Google Fonts -->
     <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@300;400;500;700&display=swap" rel="stylesheet">
     
-    <!-- Theme and App CSS -->
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/theme.css') }}">
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/app.css') }}">
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/player-controls.css') }}">
+    <!-- Consolidated Application Styles -->
+    <link rel="stylesheet" href="{{ url_for('static', filename='styles.css') }}">
     
     {% block extra_head %}{% endblock %}
 </head>
