+++ templates/search.html
@@ -2,10 +2,6 @@
 
 {% block title %}Advanced Search - Local Video Server{% endblock %}
 
-{% block extra_css %}
-<link rel="stylesheet" href="{{ url_for('static', filename='css/advanced-search.css') }}">
-{% endblock %}
-
 {% block content %}
 <div class="container">
     <!-- Search Header -->
